changed into DTO pattern
