--- src/main/java/com/kulakyokedici/kulakliksitesi/objects/security/AuthRequest.java
@@ -1,17 +1,0 @@
-package com.kulakyokedici.kulakliksitesi.objects.security;
-
-public class AuthRequest
-{
-	String username;
-	String password;
-	
-	public String getUsername() 
-	{
-		return username;
-	}
-	
-	public String getPassword() 
-	{
-		return password;
-	}
-}
