+++ src/main/java/com/kulakyokedici/kulakliksitesi/objects/security/AuthRequestDto.java
@@ -1,0 +1,17 @@
+package com.kulakyokedici.kulakliksitesi.objects.security;
+
+public class AuthRequestDto
+{
+	String username;
+	String password;
+	
+	public String getUsername() 
+	{
+		return username;
+	}
+	
+	public String getPassword() 
+	{
+		return password;
+	}
+}
